ready for more training. norm happens before converting to float16

+++ tinydas/utils.py
@@ -131,10 +131,10 @@ def reparameterize(mu: Tensor, logvar: Tensor) -> Tensor:
     return eps.mul(std).add(mu)
 
         
-def load_das_file(filename: str, dtype=dtypes.float16):
+def load_das_file(filename: str):
     """Loads a single das file in to a tuple of the data and the timestamps."""
     with h5py.File(filename, "r") as f:
-        data = Tensor(f["raw"][:], dtype=dtype, requires_grad=False).T
+        data = Tensor(f["raw"][:], requires_grad=False).T
         times = Tensor(f["timestamp"][:], requires_grad=False)
     return data, times
